Add getMetadataFromGlobalAssignment AST parser
Also factored out the metadata-related stuff into a separate file, creating some additional files in the process to avoid circular imports.
