+++ src/cli/duplicate.ts
@@ -4,10 +4,9 @@ import {
   checkIfCommandExists,
   stripName,
   kitMode,
-  stripMetadata,
-  uniq,
 } from "../core/utils.js"
 import { generate } from "@johnlindquist/kit-internal/project-name-generator"
+import { stripMetadata } from "../core/metadata"
 
 let examples = Array.from({ length: 3 })
   .map((_, i) => generate({ words: 2 }).dashed)
