+++ src/cli/new-from-protocol.ts
@@ -7,7 +7,6 @@ import { highlightJavaScript } from "../api/kit.js"
 import {
   checkIfCommandExists,
   kitMode,
-  stripMetadata,
   uniq,
 } from "../core/utils.js"
 import {
@@ -15,6 +14,7 @@ import {
   prependImport,
 } from "./lib/utils.js"
 import { generate } from "@johnlindquist/kit-internal/project-name-generator"
+import { stripMetadata } from "../core/metadata.js"
 
 let examples = Array.from({ length: 3 })
   .map((_, i) => generate({ words: 2 }).dashed)
